week06/gae/project7/requirements.txt
Flask==1.1.1
google-cloud-datastore==1.7.3